Maybe I found a problem
I found that I couldn't exit the process, even if I clicked "Exit". I've been thinking about threads, and I think it's the infinite ocr_loop that prevents daemons from being rolled out. So I added a flag bit and fixed the problem. If other users are not experiencing this problem, ignore it